Fix RootFile Unit Test: use AlertsProperties Mock
RootFile was instantiate with an autowired annotation. Bad jokes, because all integrations test execute after get the value of AlertsProperties given in this test ! So, it was no more the real value. We fix it with a AlertsProperty Mock, which is given to the RootFile Constructor. Like that, the test don't interact badly with integrations tests execute after.